Removed the need to ROSTR macro (attaching symbols to string literals should be handled in the compiler), fixed some printing warnings.

+++ lib/stack_transformation/include/list.h
@@ -26,7 +26,7 @@
 node_t(type) { \
   node_t(type)* prev; \
   node_t(type)* next; \
-  size_t act; \
+  int act; \
   type data; \
 };
